Add short-circuiting in try_union and try_intersect
Summary: This is an easy optimization: I've noticed that in `try_union` and `try_intersect`, when we find a type `ty'` in the list `tys` that is a super- resp. subtype of a given type `ty` with which we union resp. intersect, we don't stop but continue to iterate with the found super- resp. subtype.  This seems unnecessary as the list `tys` is assumed to contain no redundancy.

Reviewed By: andrewjkennedy

Differential Revision: D13635904

fbshipit-source-id: 498cee083bc3fe97de8d3162286f9dce20e9c4f4