Don't store (nested) namespace in parents, filter out empty ones
Summary:
In light of shifting error logic to "smart" constructors,
we want avoid linear-time checks on AST that go up (e.g., parents)
and eventually replace it by context/visitor that passes it down.

When encountering namespace declaration, eagerly filter it out
if it's empty so that is_in_namespace function doesn't have to
(inefficiently) search for the first non-empty one.

Reviewed By: Wilfred

Differential Revision: D13632382

fbshipit-source-id: 2961f5d32cf358505a453ab26499d38edb666c62