Enable BOLT on clang opt builds
Summary:
BOLT is currently only enabled for opt-hhvm builds. Enable it
for all opt-clang-hhvm builds too. Remove a modification in hugify made
to support a different binary layout in AArch64 as this code reference
is incompatible with BOLT. In particular, in some x86 clang builds, this
code may trigger a crash in BOLT.

Reviewed By: ottoni

Differential Revision: D13640320

fbshipit-source-id: d0f5b9713914deda33076eb89d9579ff02d7cd91