Relax restriction on reified generics and add <<__Soft>>
Summary:
For migration purposes, we're allowing people to write `reify` in a position that isn't reified. This way, `reify` can be added at all callsites of a function incrementally.

Additionally, when a definition annotates a reified type parameter with `<<__Soft>>`, it will not error on missing `reify` keyword at a use site.

Reviewed By: oulgen

Differential Revision: D13629960

fbshipit-source-id: 55118d076264521bcc6557f308c787eaf2cb9c59