Do not use declarations in scope names
Summary:
In builds with debug_types, avoid pushing to the scope stack
the name inferred from the declaration die when we have the real type
available. Look the type up and push the name of the definition. In
clang builds, we are seeing the visibility of types being wrongly
classified as "None" because the parent was an incomplete declaration
with no name, even though it was linked to the def via DW_AT_signature.

Reviewed By: ricklavoie

Differential Revision: D13646616

fbshipit-source-id: c840954a352112b5b72947d29cc0bd2a360ff9e9