Make better use of thisLocToKill, and rename it
Summary:
thisLocToKill was added to keep track of locals managed by
InitThisLoc, with the intent of removing the variable whenever
possible.

We can also use it to simply keep track of the fact that a variable
holds the $this value, which also requires tracking stack slots that
contain the $this value.

This lets us do more with Same/NSame, and other optimizations when we
know that the cell we're operating on is $this.

Reviewed By: ricklavoie

Differential Revision: D13599508

fbshipit-source-id: 264ad69f890ae3b1e10f6ed3b49207849a0dd800