Typecheck <<__SoftLateInit>> same was as <<__LateInit>>
Summary:
We (temporarily) have `<<__SoftLateInit>>` attribute to help in the migration of properties assigned MUST_PREPARE, but the type checker doesn't know about it.
This attribute should be treated the same way as `<<__LateInit>>` by the typechecker.
The difference is that it returns a constant instead of throwing an exception on access

Reviewed By: kmeht

Differential Revision: D13666326

fbshipit-source-id: ab683e5c185fc928d28ab882c5026a84f5880447