Rx verification: allow writes to arrays nested in locals.
Summary:
As discussed, have the verifier allow writes through arrays nested inside local objects (same way we already allowed writes through arrays nested in $this).
Also does a little rearranging in the verifier to prepare for mutability enforcement, but leave that as a TODO (including the unit test, which does not yet have the needed Rx\mutable annotations; the intent is to add them to the test together with adding enforcement for them).

Reviewed By: ricklavoie

Differential Revision: D13576674

fbshipit-source-id: dd140719763cf8d311b07da5e5e2597021a5aaee