Do not infer the any type when lambda has no return type annotation
Summary:
In some circumstances the type displayed by `hh_show` for a lambda involved the "any" type, displayed as `_`.
This is an unsoundness in the checking of lambdas (see tests).

Fix is simply to pass `None` for the expected type in bidirectional type checking, when there is no explicit return hint.

Reviewed By: arun11111

Differential Revision: D13692386

fbshipit-source-id: df34a257d256e2329637247a6a111af493bcd2d0