Speed-up property type-hint enforcement for APCObject
Summary:
When re-creating an object from APCObject, it is required to validate the
object's property values against their type-hints. This is because some of the
type-hints may change meaning between requests, and therefore the object may
have valid values in one request and not in another.

However, if the values conservatively satisfy their type-hints in all contexts
(regardless of what type-aliases are defined), then we can skip the checks. Do
this conservative check when creating the APCObject and set a bit if it
passes. That bit will cause us to skip the type-hint checks when re-creating the
object.

Reviewed By: markw65

Differential Revision: D13689713

fbshipit-source-id: 01cc91021c0b3efb6dd67cd0a8b4865c32844f60