Rename write_through to write_around in SharedMem.WithCache
Summary: This `write_through` function was added in {D3023663}, but I think it is a bit of a misnomer. The strategy of writing directly to a backing store without writing to a cache layer in front of it is called write-around, not write-through. The `add` function implements the write-through strategy, since it synchronously writes to both the cache and the backing store before allowing the caller to continue execution (making the existence of an alternative named `write_through` all the more confusing).

Reviewed By: pittsw

Differential Revision: D13659733

fbshipit-source-id: 1254fb941ab89e3b7785d1ec46954fe8c427337f