Improve equality check linter message for opaque enums
Summary:
Currently, this linter claims that an opaque enum can never be equal to a value of its backing type. This is clearly not true at runtime, which causes confusion (see attached task).

Instead of claiming that these values can never be equal, we should emit a more detailed message calling out the fact that the enum type is opaque, and directing the user to the functions Enum::assert and Enum::coerce, which they should use to convert the value of the backing type before comparing it to a value of the enum type.

Reviewed By: Wilfred

Differential Revision: D13676661

fbshipit-source-id: a78d98fe01e5d1848616f18f7fbcda83484b9f9a