Normalize ActRec context
Summary:
Use TObj instead of TCls in FPushCtor opcodes, as the target constructor method
gets full `$this` and not just a class context.

Define the meaning of ActRec context's isCtx as whether it corresponds to the
context of the caller. There was some confusion amongst FPushCtor opcodes,
where one opcode did explicitly setctx() while other did unctx(), so this diff
fixes it.

This context gets consumed by two places:
- return_with_context(), which tries to improve the (usually `this`) return
  type and needs to know whether it corresponds to the caller's context
- entry_state(), which sets up initial state of inlined function; we now
  setctx() this value in all cases as it is a context of the callee used while
  analyzing callee

Reviewed By: ricklavoie

Differential Revision: D13709270

fbshipit-source-id: 529ce9013a5234b197423bc05a6d9d46b7848cf2