Re-introduce array_walk and array_walk_recursive
Summary:
We need array_walk and array_walk_recursive to stick around for a bit to give open source users a chance to migrate, so add them back. There are several differences compared to the old implementation:
1) They're marked __Deprecated
2) It's in pure hack, so they don't have the magic ref behavior that caused the C++ implementation to be removed
3) If you nest an array inside itself with references, array_walk_recursive won't detect that it already visited the array and will recurse until it runs out of stack
4) This implementation of array_walk_recursive checks for hack arrays before it starts doing work and it will bail out with a warning if it detects any. The old C++ implementaion would segfault in this case (github bug #8375).

I did not add back any of the tests that were deleted together with the old implementation since, because of the warnings produced by the __Deprecated tags, they won't pass without modification anyway. I added new tests for the specific behavior we care about instead.

Reviewed By: jano

Differential Revision: D13720939

fbshipit-source-id: 36cb63f1a942b1a5a4eb033fe9b6eb5a1b6326fd