New inference: update variance of type variables after solving
Summary:
After a type variable `var` has been "solved", or bound to a type `ty`, we need
to update the variance of type variables occurring in `ty`. Suppose that
variable `var` is marked "appears covariantly", i.e. it appears (at least) in
positive positions in the type of an expression. Then when we substitute `ty`
for `var`, variables that appear positively in `ty` must now be marked as
appearing covariantly; variables that appear negatively in `ty` must now be
marked as appearing contravariantly. And the dual, if the variable `var` is marked
"appears contravariantly".

No new tests for this diff, but this change is a prerequisite for the next diff in the stack. Consider
```
class A<+T1, -T2 as T1> { ... }
$x = new A();
```
We will generate
```
$x : A<#1, #2>
-#2 <: +#1
```
Now if we decide to solve `#1 := #2`, then we need to update the variance of #2 to appear both covariantly and contravariantly.

Reviewed By: manzyuk

Differential Revision: D13750362

fbshipit-source-id: 3fd926e2511360ce746ffb650da0492cf68c9c32