Fix non-strict aliasing in PCFilter
Summary:
There was lots of type punning going on. In particular, casting m_root
to a PtrMapNode*, then modifying m_root, and assuming that the cast
pointer's m_entries field would notice the update.

gcc9 in lto mode mis-optimized some of these, so lets rewrite it to
avoid (most of) the type punning.

Reviewed By: ricklavoie

Differential Revision: D13747402

fbshipit-source-id: dee77c416f08eed6d4b141f4e87b3ce34d18e82f