Support legacy behavior of HH\BuiltinEnum in the new inference algorithm
Summary:
The subtyping relation as currently defined is not transitive.  Every enum `E` can also be view as a class, and the two types are considered equivalent with respect to subtyping (i.e., enum `E` is both a subtype and a supertype of class `E`).  `E` viewed as a class has `HH\BuiltinEnum<E>` as a parent, which gives access to the enum methods like `assert`, `getNames` etc.  However, according to the rule implemented by `simplify_subtype`, `E` viewed as an enum is never a subtype of `HH\BuiltinEnum<t>`.  In the legacy algorithm this gap doesn't matter because we fall back to unification, but with the new type inference algorithm we run into a situation where we need to assert `enum E <: HH\BuiltinEnum<v>` for some type variable `v`, which cannot be satisfied and as a result we reject code that is accepted by the legacy algorithm.

Because `HH\BuiltinEnum` is an invariant abstract base class, subtyping through the upper bound `HH\BuiltinEnum<E>` of `E` can only satisfy subtyping against `HH\BuiltinEnum<E>` itself, which is why I chose to add a separate rule instead of using the disjuction operator `|||`.

Reviewed By: andrewjkennedy

Differential Revision: D13761490

fbshipit-source-id: dc63f7ac11e8034b8e828675b9652ea694530dad