Emit VerifyParamTypeTS instead of VerifyParamType when the parameter can be reified
Summary:
This diff prepares the code base to add the ability to verify parameter type for reified generics

We do this by emitting `VerifyParamTypeTS` instead of emitting `VerifyParamType` for the parameter.
On the next diff, `VerifyParamTypeTS` will pop a type structure from the stack, runs `VerifyParamType` operation and then uses the type structure to verify the reified generics of the input.

Currently it performs identical operation to `VerifyParamType`.

Reviewed By: ricklavoie

Differential Revision: D13612514

fbshipit-source-id: 340113837eafe863607b5dadd0f7fafb1de94d2d