revise callstubs defs for aarch64 target (#8401)
Summary:
This commit https://github.com/facebook/hhvm/commit/98c4f446b683756cf7f98aef8fa9ac97d791b4c1
changed the registers presented to the allocator for the callstub Vinstr.  The comment says
it did not change the register sets for aarch64.  This is not correct.  This change caused an
internal assert to fire when it was trying to create spill space on the stack.  This sounds like
the problem seen on other hosts.

The standard regression tests were run with 6 option sets.  This brings the results back to
the normal baseline for the lab machine I use.

This fixes issue https://github.com/facebook/hhvm/issues/8396
Pull Request resolved: https://github.com/facebook/hhvm/pull/8401

Reviewed By: mofarrell

Differential Revision: D13451995

Pulled By: ricklavoie

fbshipit-source-id: 3e00089d5eaac034c8b71f6d7eb4be5b74f70d2b