New inference: spurious circular types
Summary:
We have the following situation: `tv_env` is
```
int, #34 <: +-#23
    -#34 <: #23
```
and we want to solve for `#34`. So `#34 := #23` (the upper bound), leaving
```
int, #23 <: +-#23
```
after substitution. When we come to solve `#23`, we fail with an occurs check.

Now we shouldn't end up in this situation at all: when we solve for `#34` we should remove any lower bounds on `#23` that are the variable `#34`. This generalizes to the situation (in future) where we construct an intersection from multiple upper bounds, and is precisely analogous to the optimization already performed for unions of lower bounds by D13671163.

Reviewed By: manzyuk

Differential Revision: D13784647

fbshipit-source-id: 443bb8e0a510a19e972d8aee4b7a06a596a73577