Free code.prof after retranslate-all completes
Summary:
This diff uses madvise to mark the memory region used for code.prof to be freed
by the kernel.  This is done through the treadmill to enforce that no profile
code is running anymore.

This change exposed several small issues that were causing profile code to run
or be accessed indefinitely after retranslate-all was finished, which are also
fixed here:
 1. Need to reset func prologues and func body pointers when invalidating profile translations;
 2. Need to set functions as being profiled in ProfData after emitting a body dispatch;
 3. TransKind in genFuncBodyDispatch was hard coded TransLive;
 4. Clear the list of incoming branches in the SrcDB when invalidating translations;
 5. Process all functions that were profiled in invalidateFuncsProfSrcKeys().

Reviewed By: mofarrell

Differential Revision: D13738210

fbshipit-source-id: ab1c7056ea2904015e14dd9fab8f46c867c1aaae