Expand the bounds when solving type variables
Summary:
Suppose that we have a type variable environment containing
```
t, #2 <: +#1
t <: +#2 <: #1
```
Solving for `#2` and then for `#1` we end up with `#2 := t` and `#1 := (t | #2)`.  The union in `#1` is redundant and may cause performance issues.  (See `artillery.php` below, which takes 7 minutes to type-check without changes from this diff.)  As a short term solution, we can expand the bounds when solving type variables.

(This still doesn't help if we first solve for `#1` and then for `#2`.  We could normalize unions on reading, similarly to `Typing_utils.fold_unresolved`, but we want to postpone this until we have an improved representation of unions.)

Reviewed By: andrewjkennedy

Differential Revision: D13784728

fbshipit-source-id: d7eaf7b58f181b87e68ce99e0d61c1a5a89e9403