Fix crash in debugger
Summary:
Due to problems with RequestInfo (RI) lifecycle callbacks, there are sometimes RI's in the RI instance list that that will never have requestShutdown called on them. This causes problems when the debugger tries to copy the instance list and keep it up to date via deltas using the lifecycle callbacks.

To avoid this, treat the instance list maintained on the RI side as authoritative, rather than trying to enumerate over the cached copy.

Reviewed By: ebluestein

Differential Revision: D13791095

fbshipit-source-id: 9e1747bea2903df862989096975042856725077d