Clean up printing of types in error messages
Summary:
The `Typing_print.error` function doesn't currently take an environment as parameter, which means that it is unable to expand type variables and thereby show types deeper than the top level. (For example, it just says "a nullable type" rather than anything more informative for `?int`, say.)

We add an environment parameter, and now only pass localized types to the function. Subsequent diffs will improve the actual messages that are used for types.

Reviewed By: manzyuk

Differential Revision: D13818055

fbshipit-source-id: 0b6370ac3b306918142d556ab088b0e5cbdae6fe