Default invalid mode to partial in Hack files
Summary: Previously, if a user entered an invalid mode, we'd default it to PHP. However, that would cause the legacy parser shim to suppress all parse errors in the file, including the invalid mode error! In this diff, we default an incorrect mode to partial in Hack files, which correctly displays the parse error.

Reviewed By: vassilmladenov

Differential Revision: D13810627

fbshipit-source-id: a9f092aa7d2ca870d81cf96f800e6a14f9c3d8f8