Tweak toPHPArrayIntishCast to intish cast even on PHP arrays
Summary:
Yeah it sounds pretty sad. Running SLs to see how this is on perf. At
least one unit test is affected since it constructs an array from array casting
an object with intish keys :(

Also worth mentioning this _doesn't_ change the behavior of toPHPArray [not the explicit-intish-casting-variety] since this would also affect array casts in Hack and _everywhere else_ which was not the intention

Reviewed By: periodic1236, mxw

Differential Revision: D13678796

fbshipit-source-id: 80bd793c3b8f88e5b52c508ac46921c3f0ed8c7b