Disallow non-toplevel functions before lowering
Summary:
The current approach of relying on syntax level in order to determine
if the declaration of a function (or class) is non-top-level has been
heavily criticized due to brittleness.  Instead, introduce a
notion of toplevel context **while parsing** functions and
let it be true and false for function in declaration and statement
positions, respectively, then raise error if we're type-checking
(the same conditions as before).

*Motivation*: this allows disabling non-toplevel declarations systematically,
not only when in typechecker but also when HHVM runtime option is set.

Reviewed By: pittsw

Differential Revision: D13799154

fbshipit-source-id: d4beb3a8fe0d0c8445712d0062fdc79490784fb9