Improve prop type-hint enforcement overhead for Thrift deserialization
Summary:
If the Thrift spec for a particular field implies that the field's type-hint
will always be honored, we can skip the runtime check when deserializing that
field. For example, if the spec says the field is T_STRING, and the type-hint is
string, we'll only ever deserialize strings into it, so we don't need to do the
type-hint check. This information can be calculated ahead of time when we
compile the Thrift spec.

Reviewed By: paulbiss

Differential Revision: D13774020

fbshipit-source-id: dc8f575fb3c5e9008a5e2aeefd91f4b13e43cabb