Stop implicitly coercing collections to arrays in func args.
Summary:
This coercion is left over from when we were trying to migrate from arrays to collections. We're migrating to hack arrays now, so this has outlived its usefulness. Delete the code that implements it, adjust the test.

This code happened to be the last user of Func::mustBeRef, an awful manual whitelist of builtins that were allowed to take some (or all) their arguments either by reference or by value. Delete this also.

Reviewed By: paulbiss

Differential Revision: D13793474

fbshipit-source-id: 2e8421ee117b168ede3d5ebaa1ba3c549ea49e8b