Expand jit::Type to allow an arbitrary number of type bits
Summary:
We're out of bits in `jit::Type`, so it's time to expand it.

Rather than adding a fixed number more, this diff replaces the single
`uint64_t` with an expandable bit set. Unfortunately, a surprising amount
of code relies on the existing type constants being constexpr, and
`std::bitset` does not support this.
The bulk of this diff therefore implements a simple bit set structure
that can be dropped into place where we currently use enum values.

Perf with <=64 bits *should* be identical to the existing version, as the
basic operators all optimize down to the same binary operations used now.
Will coordinate with swtaarrs to test performance with 128 bits.

Reviewed By: ricklavoie

Differential Revision: D9127872

fbshipit-source-id: 5eeb5694a1e9698094aeaf44a74e47d23369f5e3