More constraint-friendly way of type-checking the ?? operator
Summary:
At the moment, type-checking of expressions of the form `e1 ?? e2` is viewed as a special case of type-checking ternary if-expressions, so that `e1 ?? e2` becomes effectively `e1 !== null ? e1 : e2`.  Because of this, type-checking of `e1 ?? e2` triggers eager solving of the type of `e1`, which may lock the type too early and cause false type-checking positives.  This eager solving seems really unnecessary: at least for type-checking purposes, `e1 ?? e2` may be viewed as syntactic sugar for a function call `coalesce(e1, e2)`, where the signature of `coalesce` is
```
function coalesce<Tr, Ta as Tr, Tb as Tr>(?Ta, Tb): Tr;
```
This approach doesn't require eager solving of the type of `e1` and seems to work better with the constraint-based type inference algorithm.

Reviewed By: andrewjkennedy

Differential Revision: D13819445

fbshipit-source-id: da3fb4cffdd166c57fad7a3ecc11c3573cb46dda