move check_progress out of conn
Summary:
Previously, `check_progress` was a unit->unit function created inside `ClientConnect.connect` and stored in `conn`. It has the side effect of altering some mutable values `ClientConnect.progress` and `ClientConnect.progress_warning`.

I find this unsettling! I'm uneasy that the function looks like it's local and self-contained and stored inside `conn`, but in fact has the side effect of altering shared mutable variables.

This diff switches it around so that `conn` only stores the immutable `root`, and the `check_progress` function is defined immediately at top level immediately adjacent to the `progress` and `progress_warning` mutable variables which it updates.

Note: the overview is that when we call Connect, then we get an `conn.(ic,oc)` pair to let us talk directly to the server. But we need something else to let us talk to the monitor. Having `root` is the most minimal amount of information needed to talk to the monitor. That's why I think that `conn` should store `conn.root` alongside `conn.(ic,oc)` - so it can talk to both server and monitor at will.

Reviewed By: dabek

Differential Revision: D13542701

fbshipit-source-id: 30fe603e67c02324b9ab8d45b3f4baf7aeb99e59