expose more careful info about load-state failures
Summary:
Ultimately I want server to communicate with the client that we failed to load the saved state. The server needs to indicate whether it's worth doing an automatic retry, or whether a message should be exposed to the user for the user to figure out a retry action themselves.

This means I have to be extra careful that every possible source of error makes its determination about retry-ability, and takes care to have a human-readable human-actionable error message.

That's precisely what this diff does.

(I've left it as future work for a future diff to actually communicate that extra information to clientConnect and to clientLsp.)

Reviewed By: 2BitSalute

Differential Revision: D13540799

fbshipit-source-id: 613f19461c8c2be476590ecbd07358179db7782b