Consolidate nbr_procs / max_procs configuration
Summary:
This diff eliminates the GlobalConfig.nbr_procs binding and refactors logic
around configuring the max number of workers.

The max worker config can come from the local config file or command line
arguments. If both are specified, the minimum is used. The configured max
workers number is also not allowed to exceed the number of physical cores on
the machine.

This change is intented to be a pure refactor, but does change (and correct)
behavior around logging. Before this change, the max_workers number stored to
scuba did not respect the command line argument, so the running server may use
fewer workers than logged. This diff always logs the number of workers we are
actually using.

In a follow-up, I want to parameterize the shared memory with the number of
workers, to enable per-worker local storage. That requires pulling the
knowledge of how many workers we want higher in the call stack so we can pass
it to both the worker init code and the shared mem init code.

Reviewed By: gabelevi

Differential Revision: D13722880

fbshipit-source-id: c6b974f512dd1d9969041a3e4f46c36b6eaf219e