Reduce contention on the shared counter using worker-local counters
Summary:
Currently unique id generation across workers means updating an atomic variable
shared between all workers. This design leads to unnecessary contention on the
atomic which can lead to reduced parallelism.

Instead of atomically incrementing a shared counter by one each time we need an
id, this diff makes it so that workers grab a range of ids atomically which can
be consumed locally without coordination.

There is one kind of tricky point here, which is that we want to ensure the
global counter is always some multiple of the range (2048). When we hand off
from the early counter to the atomic one, we accomplish this using the same bit
fiddling used to align memory. I tweaked the ALIGNED(x) macro to reuse this bit
of code.

This doesn't seem to have a meaningful effect on either the Hack or Flow
workloads internally, but provides a reasonable first use case for worker-local
storage.

In microbenchmarks this definitely scales better (see test plan), and who
knows -- maybe having a more efficient id generation will unlock some use case
in the future.

Reviewed By: gabelevi

Differential Revision: D13742286

fbshipit-source-id: 8260e7ad67fdd5835a8fa342a31533b239c46fc3