Mention HH_IGNORE_ERROR when erroring on comments
Summary:
When a user suppresses an error with `HH_FIXME` or `HH_IGNORE_ERROR`, we error anyway if the error code is set in `ignored_fixme_codes` in `.hhconfig`.

However, the error message only mentioned `HH_FIXME`, which is confusing when the user writes `HH_IGNORE_ERROR`.

Mention both of these in the message. Note that `// UNSAFE` and `/* UNSAFE_EXPR */` are unaffected by this.

Reviewed By: arxanas

Differential Revision: D13819929

fbshipit-source-id: 45cfba68a3083366d0a6c7c947a4601751e222d1