Add \HH\Lib to qualified name for C\contains and C\contains_key in equality check linter
Summary: At some time in the past, it was sufficient to check for the qualified names `\C\contains` and `\C\contains_key` in this linter. This is now no longer true, possibly due to changes in auto-namespace-mapping logic. We now need to include "\HH\Lib" in the names we check for here.

Reviewed By: arxanas

Differential Revision: D13846678

fbshipit-source-id: 0d5ae35b1182a5ac37ea450aa2d3be95049d4436