YASOF (yet another stack overflow fix)
Summary:
The added example would create a lot of nested singleton unresolved in a nicely exponential fashion.

I figured it had to do with the way we do array_get on a Tunresolved, first mapping array_get on the type list then simply wrapping the result in a Tunresolved.
Instead, we do an approximation of a union on the list: since unioning elements of the list two by two is quadratic, we simply remove duplicates using a set (which is linear in time) then discard any singleton Tunresolved.

NB: there are other places where we could use this union_list_approx in Typing_array_access (and probably elsewhere), but as I don't know yet the effect of such a change in www, we go progressively.

Reviewed By: andrewjkennedy

Differential Revision: D13839894

fbshipit-source-id: cfb81903f1cae0bf6f3bd6f81e9f3e88b40a77d3