Unify language and mode
Summary: If a file's language is PHP, then its mode is necessarily `Mphp`. If its mode is anything else, then its language is necessarily Hack. This suggests that language and mode are not separate properties but one spectrum. This diff unifies the two so that the mode is the only property of the file.

Differential Revision: D13845478