Hide forward naming table behind API.
Summary: The forward naming table in Hack is a mapping from filenames to the symbols that occur in those files.  It's necessary for almost everything Hack does, but it's also really *really* big, so we'd like to be able to change the backing storage for it in the future.  Since the forward naming table is represented as a raw `Map`, the first step to changing its backing storage is to abstract it behind an API, which this change does.

Reviewed By: 2BitSalute, ljw1004

Differential Revision: D13705295

fbshipit-source-id: 8d3c7b65d1f79b73e2e576bc23127b6bec480376