Fix OSS build
Summary:
tbb::concurrent_hash_map doesn't work with move-only types in the OSS
build. When storing compiled Thrift specs, use a bare pointer instead of an
unique pointer. This means that they technically leak when the table is
destroyed, but its a static so it will only happen when the process exits
anyways.

Reviewed By: paulbiss

Differential Revision: D13865361