Delete RuntimeOption::EnableZendSorting
Summary:
Delete RuntimeOption::EnableZendSorting and the zend-sort code that would run instead of our normal sort implementation when you turned it on. Also delete tests that were specifically testing this different behavior (but if we can delete the option and the test still passes, keep the test).
This does not delete all the ways to reach zend-qsort.cpp, since array_multisort, array_diff, and array_intersect unconditionally went to the zend-qsort code.

Reviewed By: fredemmott

Differential Revision: D13779028

fbshipit-source-id: c5a5dfbf6cb0ea8d7db6d949dd7232d72c923393