update many tests to operate in a temporary directory
Summary:
Update many of the tests in hphp/test/zend/good/ext/standard/tests/file to
create their test files in a temporary directory rather than inside the source
tree.

This is necessary for the tests to work when the repository is checked out on
a file system that does not support hard links.  This also reduces the
performance overhead of the tests for tools like watchman and buck that
monitor the source tree for changes.

Reviewed By: markw65

Differential Revision: D13793219

fbshipit-source-id: 6f8a0be5624e7720cef4bce77270e8f1c58557e7