Emit VerifyRetTypeTS instead of VerifyRetTypeC when the return value can be reified
Summary:
This diff prepares the code base to add the ability to verify return type for reified generics

We do this by emitting `VerifyRetTypeTS` instead of emitting `VerifyRetType` for the parameter.
On the next diff, `VerifyRetTypeTS` will pop a type structure from the stack, runs `VerifyRetType` operation and then uses the type structure to verify the reified generics of the input.

Currently it performs identical operation to `VerifyRetType`.

Reviewed By: ricklavoie

Differential Revision: D13741747

fbshipit-source-id: 1df03d2d2622bee2660dc6b2ec578c1c5d4f3b08