New inference: do not check index type of covariant-key collections
Summary: Extend diff D13590142 to handle other covariant-key-type collections: `ConstMap`, `ImmMap`, `Indexish` and `KeyedContainer`. We do not want to check the type of the value used to index these collections, as covariance permits it to be treated as `arraykey`. Instead, we will implement a linter rule, or eventually make the key type invariant.

Reviewed By: manzyuk

Differential Revision: D13841953

fbshipit-source-id: 497e07779afaa4db7b32a0116d2832b99ebe981d