Correctly thread the environment when type-checking array accesses
Summary: When type-checking expressions of the form `e1[e2]`, we eagerly solve the type `ty1` of `e1` because we do a case analysis on `ty1` (to ensure, in particular, that it supports indexing).  However, in all the cases, we were erroneously passing the original environment in which `ty1` was possibly still unsolved, which was causing some false positives.

Reviewed By: andrewjkennedy

Differential Revision: D13877837

fbshipit-source-id: 3dbc897300eb3efd9fa5f64326e3e7935df3e985