remove exec_with_replacement_env
Summary:
Oh blast. I spent ages writing a summary, but my editor has lost it.

Long story short. We had a function "exec_with_replacement_env" which claimed to execute a process with a replaced environment. Its comments said this is almost never the right thing to do. Its actual behavior at runtime was either (1) augment the environment if no cwd was specified, (2) replace the environment if a cwd was specified, (3) leave the environment unchanged if a cwd was specified and you were using ocaml <= 4.05 and you were on Windows.

I've replaced it with a function "exec_with_augmented_env" which is more useful, and I made sure it only ever uses the "no-cwd" path which does indeed augment it correctly.

I updated a test which had been broken for many months, probably since April 2018.

The only place this functionality is ever used is in hg.ml where it adds `HGPLAIN=1` into the environment to make sure it can parse the output properly.

Reviewed By: losvald

Differential Revision: D13865451

fbshipit-source-id: 6d965fc98f01fdab3db217f6b03c95d47573d8f5